backend/requirements.txt
absl-py==2.2.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
autopep8==2.3.2
awscli==1.40.34
bcrypt==4.2.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
blis==1.3.0
botocore==1.38.35
cachetools==5.5.2
catalogue==2.0.10
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-rest-framework==0.1.0
django-silk==5.3.2
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
docutils==0.19
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fpdf2==2.8.3
fsspec==2025.3.2
gast==0.6.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
gprof2dot==2024.6.6
grpcio==1.72.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
Keras-Applications==1.0.8
kiwisolver==1.4.8
kombu==5.4.2
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1

mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.12
namex==0.0.9
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
patsy==1.0.1
pdfkit==1.0.0
pillow==11.1.0
platformdirs==4.3.8
preshed==3.0.9
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.5
python3-openid==3.2.0
pytz==2025.1

PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.7.2
s3transfer==0.13.0
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
thinc==8.3.6
threadpoolctl==3.5.0
tk==0.1.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.31.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.1
daphne==4.1.0

whitenoise==6.7.0

django-allauth==0.63.0